requirements.txt
Django==1.10.5
channels==1.0.3
python-twitter==3.2.1
